TIP-221: Move ProductTemplateBuilder
