+++ spec/Pim/Bundle/CatalogBundle/EventSubscriber/ProductTemplateAttributeSubscriberSpec.php
@@ -6,7 +6,7 @@ use Akeneo\Component\StorageUtils\Event\RemoveEvent;
 use Akeneo\Component\StorageUtils\StorageEvents;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Builder\ProductTemplateBuilderInterface;
+use Pim\Component\Catalog\Builder\ProductTemplateBuilderInterface;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ProductTemplateInterface;
 use Pim\Component\Catalog\Repository\ProductTemplateRepositoryInterface;
