+++ src/Pim/Bundle/CatalogBundle/Resources/config/builders.yml
@@ -1,6 +1,6 @@
 parameters:
     pim_catalog.builder.product.class:          Pim\Component\Catalog\Builder\ProductBuilder
-    pim_catalog.builder.product_template.class: Pim\Bundle\CatalogBundle\Builder\ProductTemplateBuilder
+    pim_catalog.builder.product_template.class: Pim\Component\Catalog\Builder\ProductTemplateBuilder
     pim_catalog.builder.choices.class:          Pim\Bundle\CatalogBundle\Builder\ChoicesBuilder
 
 services:
